initial, draw background
