+++ example/old/index.ts
@@ -1,0 +1,2 @@
+export * from "./types";
+export * from "./graph-view-old";
