+++ example/old/old-example/index.css
@@ -1,0 +1,59 @@
+html, body {
+  width: 100%;
+  height: 100%;
+
+  margin: 0;
+  padding: 0;
+
+  font-family: sans-serif;
+  font-size: 16px;
+}
+
+#graph {
+  width: 100%;
+  height: 100%;
+}
+
+#container {
+  width: 100%;
+  height: 100%;
+  position: relative;
+}
+
+#hints {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  margin: 1em;
+  font-size: 0.9em;
+  color: #718096
+}
+
+#buttons {
+  position: absolute;
+  top: 0;
+  left: 0;
+  margin: 1em;
+}
+
+.generate-input {
+  margin-top: 1em;
+}
+
+#generate-textbox {
+  padding: 0.5em;
+  width: 7.5em;
+  text-align: right;
+}
+
+button {
+  padding: 0.5em;
+  cursor: pointer;
+}
+
+.zoom-input {
+}
+
+.info {
+  margin-top: 1em;
+}
