+++ src/graph-view.ts
@@ -1,120 +1,139 @@
-import { GEState } from "./state";
-import { GEGraphRenderer } from "./graph-renderer";
-import { GEEventHandler } from "./event-handler";
-import { GENode, GEEdge, GEViewOptionsParams } from "./types";
+import { drawBackground } from "./graph-renderer";
 
-export class GEView {
+export type NodeShape = {
+  render: (ctx: CanvasRenderingContext2D) => void;
+  intersectionPoints: [number, number][];
+};
+
+export type EdgeShape = {
+  render: (ctx: CanvasRenderingContext2D) => void;
+};
+
+export type GraphNode = {
+  x: number;
+  y: number;
+  shape: EdgeShape;
+};
+
+export type GraphEdge = {
+  source: GraphNode;
+  target: GraphNode;
+  shape: NodeShape;
+};
+
+export class GraphView<Node extends GraphNode, Edge extends GraphEdge> {
   readonly canvas: HTMLCanvasElement;
+  readonly ctx: CanvasRenderingContext2D;
 
-  private _state: GEState;
-  private _renderer: GEGraphRenderer;
-  private _eventHandler: GEEventHandler;
+  transform: [number, number, number] = [1, 0, 0]; // [scale, tx, ty]
+  nodes: Node[] = [];
+  edges: Edge[] = [];
+  hoveredNode: Node | undefined = undefined;
+  hoveredEdge: Edge | undefined = undefined;
+  pointerPos: [number, number] = [0, 0];
+  movingNode: Node | undefined = undefined;
+  moveNodePos: [number, number] = [0, 0];
+  isCreatingEdge = false;
+  dragLineSourceNode: Node | undefined = undefined;
+  dragLineTargetPos: [number, number] = [0, 0];
 
-  constructor(options?: GEViewOptionsParams) {
+  private isDrawing = false;
+  private boundingRect: DOMRect;
+
+  constructor(container: HTMLElement) {
     this.canvas = document.createElement("canvas");
-    this._state = new GEState();
-    this._renderer = new GEGraphRenderer(this._state, this.canvas);
-    this._eventHandler = new GEEventHandler(
-      this._state,
-      this.canvas,
-      this._renderer
-    );
 
-    if (options) {
-      this._state.setOptions(options);
+    const ctx = this.canvas.getContext("2d", { alpha: false });
+
+    if (!ctx) {
+      throw "Canvas is not supported in your browser.";
     }
-  }
 
-  getTranslateX(): number {
-    return this._state.translateX;
-  }
-
-  getTanslateY(): number {
-    return this._state.translateY;
-  }
-  getScale(): number {
-    return this._state.scale;
-  }
-
-  zoomTo(value: number): void {
-    const { width, height } = this.canvas;
-    const { translateX, translateY, scale } = this._state;
-
-    const centerX = (width * 0.5 - translateX) / scale;
-    const centerY = (height * 0.5 - translateY) / scale;
-
-    this._state.zoomTo(value, centerX, centerY);
-
-    this.requestDraw();
-  }
-
-  init(container: HTMLElement): void {
-    container.appendChild(this.canvas);
+    this.ctx = ctx;
 
     this.canvas.textContent = "Canvas is not supported in your browser.";
     this.canvas.width = container.clientWidth;
     this.canvas.height = container.clientHeight;
+    this.boundingRect = this.canvas.getBoundingClientRect();
 
-    this._state.setBoundingRect(this.canvas);
-    this._eventHandler.init();
+    container.appendChild(this.canvas);
+
     this.requestDraw();
   }
 
   destroy(): void {
-    this._eventHandler.destroy();
+    // this._eventHandler.destroy();
   }
 
   requestDraw(): void {
-    this._renderer.requestDraw();
+    if (!this.isDrawing) {
+      requestAnimationFrame(this.draw);
+    }
+
+    this.isDrawing = true;
   }
+
+  draw = () => {
+    this.isDrawing = false;
+
+    drawBackground(this);
+  };
 
   resize(width: number, height: number): void {
     this.canvas.width = width;
     this.canvas.height = height;
 
-    this._state.setBoundingRect(this.canvas);
+    this.boundingRect = this.canvas.getBoundingClientRect();
 
     this.requestDraw();
   }
 
-  setData(nodes: GENode[], edges: GEEdge[]): void {
-    this._state.setData(nodes, edges);
-    this.requestDraw();
-  }
+  zoomTo(value: number, viewX?: number, viewY?: number): void {
+    const { width, height } = this.canvas;
+    const [scale, translateX, translateY] = this.transform;
 
-  setOptions(options: GEViewOptionsParams): void {
-    this._state.setOptions(options);
-    this.requestDraw();
-  }
+    viewX = viewX || (width * 0.5 - translateX) / scale;
+    viewY = viewY || (height * 0.5 - translateY) / scale;
 
-  setSelection(node: GENode | undefined, edge: GEEdge | undefined): void {
-    if (node && edge) return;
+    const newScale = Math.min(1000.0, Math.max(0.1, value));
 
-    this._state.selectedNode = node;
-    this._state.selectedEdge = edge;
+    const deltaScale = newScale - scale;
+    const offsetX = -(viewX * deltaScale);
+    const offsetY = -(viewY * deltaScale);
 
-    this.requestDraw();
-  }
-
-  setSelectedNode(node: GENode | undefined): void {
-    this._state.selectedNode = node;
-    this._state.selectedEdge = undefined;
+    this.transform[0] += deltaScale;
+    this.transform[1] += offsetX;
+    this.transform[2] += offsetY;
 
     this.requestDraw();
   }
 
-  getSelectedNode(): GENode | undefined {
-    return this._state.selectedNode;
+  setViewPosFromWindowPos(
+    out: [number, number],
+    windowX: number,
+    windowY: number
+  ) {
+    const { left, top } = this.boundingRect;
+    const [scale, translateX, translateY] = this.transform;
+
+    out[0] = (windowX - left - translateX) / scale;
+    out[1] = (windowY - top - translateY) / scale;
   }
 
-  setSelectedEdge(edge: GEEdge | undefined): void {
-    this._state.selectedNode = undefined;
-    this._state.selectedEdge = edge;
+  setViewPosFromCanvasPos(
+    out: [number, number],
+    canvasX: number,
+    canvasY: number
+  ) {
+    const [scale, translateX, translateY] = this.transform;
 
-    this.requestDraw();
+    out[0] = (canvasX - translateX) / scale;
+    out[1] = (canvasY - translateY) / scale;
   }
+}
 
-  getSelectedEdge(): GEEdge | undefined {
-    return this._state.selectedEdge;
-  }
+export function createGraphView<Node extends GraphNode, Edge extends GraphEdge>(
+  container: HTMLElement
+) {
+  return new GraphView<Node, Edge>(container);
 }
